Refactor component into individual files
